chore: fix clippy lints from newer stable in vendored core
- drop redundant & in six format!/warn!/trace! args
- iter() -> values() in the discovery subscriber loop
- drop pointless u64 widen/narrow in a battery test

+++ src/camera_tasks.rs
@@ -944,7 +944,7 @@ mod tests {
 		let _ = task.await;
 
 		assert!(
-			call.load(Ordering::Acquire) > u64::from(BATTERY_MAX_UNSUPPORTED) as u32 + 1,
+			call.load(Ordering::Acquire) > BATTERY_MAX_UNSUPPORTED + 1,
 			"test must actually drive more refusals than the streak limit"
 		);
 		assert!(
